feat: in de artefact '@dom-lib/elements' naast 'vl-elements.css' ook een 'vl-elements.css.js' aanbieden

+++ web-components/post-process-libs.mjs
@@ -3,23 +3,25 @@ import fs from 'fs-extra';
 
 const wrapAsStylesObject = (content) => `const styles = \`${content}\`; export default styles;`;
 
-const wrapCssInJs = (directoryToSearch) => {
+const wrapCssInJs = (directoryToSearch, jsSuffix, removeOriginal) => {
     const pattern = '.css';
     fs.readdirSync(directoryToSearch).forEach((file) => {
         const filePath = path.resolve(directoryToSearch, file);
         const stat = fs.statSync(filePath);
         if (stat.isDirectory()) {
-            wrapCssInJs(filePath);
+            wrapCssInJs(filePath, jsSuffix, removeOriginal);
         }
         if (stat.isFile() && filePath.endsWith(pattern)) {
-            const scssJsFile = filePath.replace(/.css$/, '.scss.js');
-            fs.readFile(filePath, 'utf8',(err, data) => {
+            const scssJsFile = filePath.replace(/.css$/, jsSuffix);
+            fs.readFile(filePath, 'utf8', (err, data) => {
                 // replace one \ with two \\ to escape the escape character
                 data = data.replace(/\\/g, '\\\\');
                 fs.outputFile(scssJsFile, wrapAsStylesObject(data));
                 console.log('processed css file to:', scssJsFile);
-                // remove the original files, keep the map file
-                fs.unlink(filePath);
+                if (removeOriginal) {
+                    // remove the original file, keep the map file
+                    fs.unlink(filePath);
+                }
             });
         }
     });
@@ -43,19 +45,20 @@ const copySources = (directoryToSearch, directoryCopyTo, pattern) => {
 };
 
 // post process elements
+wrapCssInJs('dist/libs/elements/src', '.css.js', false);
 copySources('libs/elements/src', 'dist/libs/elements/src', '.lib.js');
 
 // post process components
-wrapCssInJs('dist/libs/components/src');
+wrapCssInJs('dist/libs/components/src', '.scss.js', true);
 copySources('libs/components/src', 'dist/libs/components/src', '.lib.js');
 
 // post process sections
-wrapCssInJs('dist/libs/sections/src');
+wrapCssInJs('dist/libs/sections/src', '.scss.js', true);
 copySources('libs/sections/src', 'dist/libs/sections/src', '.lib.js');
 
 // post process test-support
 copySources('libs/support/test-support/src', 'dist/libs/support/test-support/src', '.js');
 
 // post process map
-wrapCssInJs('dist/libs/map/src');
+wrapCssInJs('dist/libs/map/src', '.scss.js', true);
 copySources('libs/map/src', 'dist/libs/map/src', '.js');
